+++ app/routes/jobs/$id/stdout.tsx
