+++ remix.config.js
@@ -7,6 +7,5 @@ module.exports = {
   // publicPath: "/build/",
   future: {
     unstable_tailwind: true,
-    v2_routeConvention: true,
   },
 };
